+++ application/controllers/admin/import.php
@@ -16,6 +16,7 @@ if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
 Class Import extends Bancha_Controller
 {
+	public $adapters = array();
 
 	public function __construct()
 	{
@@ -26,6 +27,13 @@ Class Import extends Bancha_Controller
 	    $this->view->base = 'admin/';
 
 	    $this->auth->needs_login();
+
+	    $this->adapters = array(
+	    	'csv'		=> 'CSV',
+	    	'wordpress'	=> 'Wordpress'
+		);
+
+		$this->view->set('adapters', $this->adapters);
 
 	}
 
@@ -43,5 +51,57 @@ Class Import extends Bancha_Controller
 		
 		$this->view->render_layout('import/records');
 	}
+
+	public function step($which = 1) 
+	{
+		$adapter_type = $this->input->post('adapter_type');
+
+		if (!in_array($adapter_type, array_keys($this->adapters)))
+		{
+			show_error(_('Adapter not found.'));
+		}
+
+		if (!isset($_FILES['records']) || !$_FILES['records']['tmp_name'])
+		{
+			$this->view->message('warning', _('You must upload a file.'));
+			$this->index();
+			return;
+		} else {
+			$file = $_FILES['records'];
+		}
+
+		$contents = file_get_contents($file['tmp_name']);
+		$this->load->adapter($adapter_type);
+
+		if (!in_array($file['type'], $this->adapter->mimes))
+		{
+			$this->view->message('warning', _('The mime-type of the file is not allowed by this adapter.'));
+			$this->index();
+			return;
+		}
+
+		$autosave = TRUE;
+
+		$records = $this->adapter->parse_stream($contents, TRUE, $this->input->post('type_id'), $autosave);
+		
+		if (count($records) && !$autosave)
+		{
+			foreach ($records as $record)
+			{
+				if ($this->records->save($record))
+				{
+					$saved++;
+				}
+			}
+		} else {
+			$saved = count($records);
+		}
+
+		$this->view->set('records', $records);
+		$this->view->message('success', $this->lang->_trans('%c records have been imported.', array('c' => $saved)));
+
+		$this->view->render_layout('import/report');
+	}
+
 
 }